Move cookie parsing into dedicated class
By moving this code into a dedicated class we prepare for future changes required by RFC's, and don't grow the `ServerRequest` class beyond what it purpose it.
